Move tests back to `main_test.py`
